v5 -> v6 migration impl.
