+++ modules/migrator/src/main/resources/application.conf
@@ -6,6 +6,8 @@ db.pass = "1234"
 db.pass = ${?DB_PASS}
 db.cp-size = 8
 db.cp-size = ${?MIGRATOR_DB_CP_SIZE}
+network-prefix = 0
+network-prefix = ${?NETWORK_PREFIX}
 migration-id = ${?MIGRATION_ID}
 offset = 0
 offset = ${?MIGRATION_OFFSET}
