+++ modules/migrator/src/main/scala/org/ergoplatform/explorer/migration/Application.scala
@@ -2,8 +2,10 @@ package org.ergoplatform.explorer.migration
 
 import cats.effect.{ExitCode, IO, IOApp, Resource}
 import doobie.util.transactor.Transactor
+import org.ergoplatform.ErgoAddressEncoder
 import org.ergoplatform.explorer.db.DoobieTrans
-import org.ergoplatform.explorer.migration.migrations.RegistersMigration
+import org.ergoplatform.explorer.migration.configs.{AssetsMigrationConfig, MigrationConfig, RegistersMigrationConfig}
+import org.ergoplatform.explorer.migration.migrations.{AssetsMigration, RegistersMigration}
 import pureconfig.generic.auto._
 
 import scala.concurrent.duration._
@@ -11,11 +13,10 @@ import scala.concurrent.duration._
 object Application extends IOApp {
 
   def run(args: List[String]): IO[ExitCode] =
-    resources(args.headOption).use {
-      case (xa, conf) =>
-        val migrations = makeMigrations(conf.offset, xa)
-        migrations(conf.migrationId)
-          .as(ExitCode.Success)
+    resources(args.headOption).use { case (xa, conf) =>
+      val migrations = makeMigrations(conf.offset, xa)(conf.addressEncoder)
+      migrations(conf.migrationId)
+        .as(ExitCode.Success)
     }
 
   private def resources(configPathOpt: Option[String]) =
@@ -24,6 +25,9 @@ object Application extends IOApp {
       xa       <- DoobieTrans[IO]("Migrator", settings.db)
     } yield xa -> settings
 
-  private def makeMigrations(offset: Int, xa: Transactor[IO]) =
-    Map("v4v5" -> RegistersMigration(RegistersMigrationConfig(batchSize = 1000, interval = 500.millis, offset), xa))
+  private def makeMigrations(offset: Int, xa: Transactor[IO])(implicit e: ErgoAddressEncoder) =
+    Map(
+      "v4v5" -> RegistersMigration(RegistersMigrationConfig(batchSize = 1000, interval = 500.millis, offset), xa),
+      "v5v6" -> AssetsMigration(AssetsMigrationConfig(batchSize = 1000, interval = 500.millis, offset), xa)
+    )
 }
