+++ modules/migrator/src/main/scala/org/ergoplatform/explorer/migration/configs/AssetsMigrationConfig.scala
@@ -1,0 +1,5 @@
+package org.ergoplatform.explorer.migration.configs
+
+import scala.concurrent.duration.FiniteDuration
+
+case class AssetsMigrationConfig(batchSize: Int, interval: FiniteDuration, offset: Int)
